Changed return from ModificationService to Reactive stream
This should make it a lot more flexible and efficient processing the
modified items.
I did have a quick look at doing this with the Java 9 native reactive
streams (flow api), but couldn't see how to simply do a blocking
collection the way RxJava does. RxJava still seems to do more so fell
back on that